refactor(frontend): convert `auth` to TypeScript (#78)
