+++ frontend/src/auth/use-user-info.ts
@@ -1,4 +1,4 @@
-import { useGetUserInfoQuery } from "../features/api"
+import { useGetUserInfoQuery } from "./auth.api"
 
 const useUserInfo = () => {
   const { data, isLoading, isUninitialized, isError, isFetching } =
