--- frontend/src/features/api/index.ts
@@ -1,2 +1,0 @@
-export { authApi, useGetUserInfoQuery } from "./authApi"
-export { metadataApi, useGetProposalQuery } from "./metadataApi"
