--- frontend/src/features/auth/index.ts
@@ -1,2 +1,0 @@
-export { selectUserInfoResult, selectUserFullName } from "./authSlice"
-export { login, logout } from "./authActions"
